feat: improve code - add new models cow-order and stack order - add fetcher to handle fetching data - add filters for filtering data - delete old/unused components (StacksTable, CellWrapper,...)
