+++ packages/app/models/stack-order/index.ts
@@ -1,4 +1,3 @@
-export * from "./order";
 export * from "./types";
-export * from "./filters";
+export * from "./stack-order";
 export * from "./fetcher";
