+++ packages/app/models/stack-order/types.ts
@@ -1,0 +1,6 @@
+import { Order } from "@/models/order";
+import { OrderMetaData } from "@cowprotocol/cow-sdk";
+
+export interface StackOrder extends Order {
+  cowData?: OrderMetaData[];
+}
